backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
motor>=3.3.0
pymongo>=4.6.0
openai>=1.0.0
